requirements.txt
asgiref==3.5.2
certifi
cfgv==3.3.1
distlib==0.3.6
Django==4.0
django-extensions==3.2.1
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==15.2.0
filelock==3.8.0
identify==2.5.8
nodeenv==1.7.0
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
virtualenv==20.16.6
